Add support for semi-automated cluster bootstrap.
This adds support for configuring nodes differently based on a hiera
parameter ('mariadb::cluster::build_stage'). Options are 'standalone',
which configures the node with no clustering support; 'peer', which
configures the node with clustering support and with the cluster address
set to point at the cluster peers; and 'bootstrap' which configures the
node to bootstrap the cluster, with the cluster address set to
'gcomm://'.

The cluster bootstrap is thus a three stage process, with the staging
driven by changes to the node hiera:

1) pick a node as bootstrap, set the build_stage value to 'bootstrap',
and run puppet, possibly multiple times until it stabilises and the
mariadb service is up and listening

2) on the other nodes, set the build_stage value to 'peer' and run
puppeton each nodes, possibly multiple times until it stabilised and the
mariadb service is up and has joined the cluster

3) on the bootstrap node, set the build_stage value to 'peer' and run
puppet to update the cluster address and restart mariadb with the new
value