Add the root password to the debian.cnf file.
This is necessary to keep things working when the root account is
configured to use password auth rather than unix_socket auth. Without
this the package maintenance scripts will fail in the password auth
case, which is likely to be the default for a 10.5 node which has been
upgraded from previous releases.