Add a standalone "clustercheck" script.
This supports the use case where the server is running in standalone
mode, but the clients are connecting via an haproxy (as is commonly the
case with clusters).

While this may seem a strange configuration, it is one that can occurr
if there are issues with clustering that require moving temporarily to a
standalone mode while the issues are worked on.

Note that this can cause issues if more than one node in the cluster is
running at a time, as any additional standalone nodes will also appear
to be up to the haproxy status checks - make sure that the other nodes
in the cluster are shut down, or in a mode other than standalone in
order to avoid this.