Add support for configuring a client-side haproxy instance.
This creates a local haproxy instance which can be used to manage access
to the cluster. In this mode a client connects to 127.0.0.1:3306 and
haproxy forwards connections through to the cluster, either using a
manually specified primary or randomly selecting one of the cluster
nodes to use as the primary. By default haproxy is configured to use the
cluster status check service configured by this module (see
mariadb::cluster::status); this can be disabled if required.