Rework cluster check scripts.
This was forced by the fact that the old scripts were specifying the
'--port' option on the command line, which with MariaDB 10.6 forces a
TCP connection. The fix is to remove the --port, and explicitly specify
--protocol=SOCKET - this ensures there's no room for confusion.

Since the scripts were being changed I also went in and reworked them to
be neater and a little better documented, and reduced the privileges
available to the cluster check user even further (no need for anything
beyond USAGE for these checks).