Release v1.3.1.
This is very important for anyone running MariaDB 10.6, but for earlier
versions it's not so critical.

This is because older versions of the clustercheck script (both the
cluster and standalone versions) no longer work after the MariaDB 10.6
client release, due to changes in the handling of the `--port=` option
in combination with `--host=localhost`.